rearrange files in packages
