+++ pkg/cli/targets.go
@@ -3,6 +3,7 @@ package cli
 import (
 	"encoding/json"
 	"fmt"
+	"github.com/mattfenwick/cyclonus/pkg/explainer"
 	"github.com/mattfenwick/cyclonus/pkg/matcher"
 	"github.com/mattfenwick/cyclonus/pkg/utils"
 	"github.com/spf13/cobra"
@@ -72,12 +73,12 @@ func RunQueryTargetsCommand(args *QueryTargetsArgs) {
 		ingressValue := true
 		ingressTargets := explainedPolicies.TargetsApplyingToPod(ingressValue, pod.Namespace, pod.Labels)
 		for _, t := range ingressTargets {
-			fmt.Printf("    %s\n", strings.Join(matcher.ExplainTarget(t, ingressValue), "\n"))
+			fmt.Printf("    %s\n", strings.Join(explainer.ExplainTarget(t, ingressValue), "\n"))
 		}
 		// combine all the ingress targets for combined connectivity
 		combinedIngressTarget := matcher.CombineTargetsIgnoringPrimaryKey(pod.Namespace, metav1.LabelSelector{MatchLabels: pod.Labels}, ingressTargets)
 		if combinedIngressTarget != nil {
-			fmt.Printf("    combined ingress:\n%s\n\n", strings.Join(matcher.ExplainTarget(combinedIngressTarget, ingressValue), "\n"))
+			fmt.Printf("    combined ingress:\n%s\n\n", strings.Join(explainer.ExplainTarget(combinedIngressTarget, ingressValue), "\n"))
 		} else {
 			fmt.Println("    combined ingress: none")
 		}
@@ -87,12 +88,12 @@ func RunQueryTargetsCommand(args *QueryTargetsArgs) {
 		egressValue := false
 		egressTargets := explainedPolicies.TargetsApplyingToPod(egressValue, pod.Namespace, pod.Labels)
 		for _, t := range egressTargets {
-			fmt.Printf("    %s\n", strings.Join(matcher.ExplainTarget(t, egressValue), "\n"))
+			fmt.Printf("    %s\n", strings.Join(explainer.ExplainTarget(t, egressValue), "\n"))
 		}
 		// combine all the egress targets for combined connectivity
 		combinedEgressTarget := matcher.CombineTargetsIgnoringPrimaryKey(pod.Namespace, metav1.LabelSelector{MatchLabels: pod.Labels}, egressTargets)
 		if combinedEgressTarget != nil {
-			fmt.Printf("    combined egress:\n%s\n\n", strings.Join(matcher.ExplainTarget(combinedEgressTarget, egressValue), "\n"))
+			fmt.Printf("    combined egress:\n%s\n\n", strings.Join(explainer.ExplainTarget(combinedEgressTarget, egressValue), "\n"))
 		} else {
 			fmt.Println("    combined egress: none")
 		}
